requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-cors-headers==2.4.0
django-rest-framework-docs==0.1.7
django-sendfile==0.3.11
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-writable-nested==0.5.1
drf-yasg==1.11.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpickle==1.0
MarkupSafe==1.1.0
Pillow==5.3.0
PyJWT==1.6.4
pytz==2018.5
requests==2.20.1
ruamel.yaml==0.15.77
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
uuid==1.30
uwsgi==2.0.17